split up actina api module
